requirements.txt
Flask
Flask-SocketIO
requests
gitpython
gensim
scikit-learn
pandas
plotly
